backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
google-generativeai
firebase-admin
httpx
python-dotenv
pydantic
git+https://github.com/internetarchive/openlibrary-client.git
